PR #18251: Fixed "reset_state" of R2Score metric
Imported from GitHub PR https://github.com/keras-team/keras/pull/18251

"reset_state" method of R2Score metric throws an error:
`ValueError: Tensor conversion requested dtype int32 for Tensor with dtype float32: <tf.Tensor: shape=(), dtype=float32, numpy=0.0>`
Quick fix is to specify original "dtype" of variable when assigning it to zeros.
Copybara import of the project:

--
fd6692a9fdc8b4969e49788c783e1c0f2be847a5 by Pavel Dyakov <pl.dyakov@gmail.com>:

Fixed "reset_state" of R2Score metric

Merging this change closes #18251

FUTURE_COPYBARA_INTEGRATE_REVIEW=https://github.com/keras-team/keras/pull/18251 from pdyakov:patch-1 fd6692a9fdc8b4969e49788c783e1c0f2be847a5
PiperOrigin-RevId: 544417576

+++ keras/metrics/regression_metrics.py
@@ -598,7 +598,7 @@ class R2Score(base_metric.Metric):
 
     def reset_state(self):
         for v in self.variables:
-            v.assign(tf.zeros(v.shape))
+            v.assign(tf.zeros(v.shape, dtype=v.dtype))
 
     def get_config(self):
         config = {
